Bug 2009146 - upgrade NSS to 7dbff6ce7aace164b95676e48b91ec27ee415f73. r=nss-reviewers,jschanck UPGRADE_NSS_RELEASE
Differential Revision: https://phabricator.services.mozilla.com/D282858

UltraBlame original commit: a98694eb4685b1ab689278aaffc98188634992a8

+++ security/nss/lib/freebl/gcm.c
@@ -436,7 +436,7 @@ gcmHash_Update(gcmHashContext *ghash, const unsigned char *buf,
     unsigned int blocks;
     SECStatus rv;
 
-    ghash->cLen += (len * PR_BITS_PER_BYTE);
+    ghash->cLen += ((uint64_t)len * PR_BITS_PER_BYTE);
